feat: move HeroBannerDescription component to folder and add the test
