+++ src/components/molecules/HeroBanner.tsx
@@ -2,7 +2,7 @@ import { FC } from "react";
 
 import { Grid, Card, Box } from "@mui/material";
 import { Link } from "react-router-dom";
-import HeroBannerDescription from "../atoms/HeroBannerDescription";
+import HeroBannerDescription from "../atoms/HeroBannerDescription/HeroBannerDescription";
 import HeroBannerTags from "../atoms/HeroBannerTags";
 import HeroBannerTitle from "../atoms/HeroBannerTitle";
 import HeroBannerRating from "../atoms/HeroBannerRating";
